2019-12-10 Commit ( Create Menu & Portfolio Menu modifying)
Once cloned, I need you to add some fields to the Create menu. A “URL” field needs to be added as well as a “Decimals” field.

Once that’s completed, under the “Portfolio” menu, if you select a token you’ve created, and press the “...” menu, another option will appear called “Pay Dividends”. This is currently blocked to normal users on pitico.cash, and I’d like it available and working on our project. The goal of it is for users to be able to send BCH to every address that owns some of the selected token, proportional to how much they own out of the total supply.